move sell order function to utils
